+++ sematic/ui/src/pipelines/LogPanel.tsx
@@ -1,7 +1,7 @@
 import { Alert, Box, TextField } from "@mui/material";
 import { useCallback, useState } from "react";
 import { Run } from "../Models";
-import { Exception, ExternalException } from "./Exception";
+import { Exception, ExternalException } from "../components/Exception";
 import ScrollingLogView from "./ScrollingLogView";
 
 export default function LogPanel(props: { run: Run }) {
