+++ sematic/ui/src/pipelines/graph/ArtifactNode.tsx
